backend/face_service/requirements.txt
flask==3.1.0
deepface>=0.0.93
Pillow>=11.0.0
opencv-python-headless>=4.8.0
numpy>=1.24.0
tf-keras>=2.16.0
torch
torchvision
